fix(Pointer): prevent bezier pointer joint clipping on down cast
The Bezier Pointer would occasionally flicker between the valid/invalid
state specifically when hovering over a terrain due to the way the
forward cast joint position could be placed where when the down cast
happened it was already clipping through the target object meaning no
ray collision would occur.

The joint position already had an offset to move the joint back a
slight amount across the beam forward length, but this adjustment
also needed applying to `y` of the beam to ensure it always started
just above the target below.