fix(Interaction): replace force drop enumerator for list index access
The `ForceStopAllInteractions` method previously used an enumerator
to traverse the `touchingObjects` list. However, this list could
change during the course of the call therefore changing the enumerator
state breaking the loop and crashing the program.

The solution is to look through the `touchingObjects` list using a
simple for loop with index access to the list. Therefore if something
changes, it shouldn't break the enumerator as it's not being used.