feat(AdaptiveQuality): adjust render target resolution
The render target resolution is adjusted to always be as high as
sensible. It will be scaled down if there's no need for a huge
resolution and scaled up whenever the current VR framerate allows.
Scaling down the resolution frees up VRAM but the cost of changing the
resolution is very high. Therefore this feature is disabled by default
and can be toggled on via `scaleRenderTargetResolution`.