feat(Pointer): options to reorient and rescale the pointer cursor
Add to VRTK_SimplePointer an option to rotate the pointer cursor to
match the normal of the target surface (pointerCursorMatchTargetNormal).

Add to VRTK_SimplePointer an option to scale the pointer cursor
proportionally to the distance from this game object if this is true
(pointerCursorRescaledAlongDistance).

Both these options are useful when using the simple pointer as a gaze
pointer with a custom pointer without a spherical symmetry.