feat(Pointer): check headset out of bounds on play area cursor
A new setting on the Play Area Cursor script has been added that
determines if the headset is outside the bounds of the defined
cursor play space and if it is outside then it treats this as a
play area collision.

This is useful if preventing teleporting if the user isn't standing
within the defined play space is required.