fix(Interaction): objectHighlighter not initialized in ToggleHighlight
Add a check on objectHighlighter in
VRTK_InteractableObject.ToggleHighlight.

Set objectHighlighter to null after it is destroyed in OnDisable().

With a `NetworkManager` active and if a `NetworkIdentity` component is
added to the same game object holding `VRTK_InteractableObject`, then
`VRTK_InteractableObject.objectHighlighter` is destroyed before the
object is touched.
This is a workaround to ensure objectHighlighter is initialized anyway.