feat(Controller): add headset controller awareness
The Headset Controller Aware script provides the ability to have the
headset aware of where the controllers are in relation to the headset.

Two key elements are provided:

 * The headset can track if the controllers are being obscured from a
 direct path between the headset and controller.
 * The headset can know when it is looking at a controller.

If a controller is obscured from the headset then an event is emitted
as well as when the controller is no longer obscured.

Also, when the headset looks (glances) at a controller then an event
is emitted.

The Controller Tooltips prefab has been updated to only show the
controller tooltips when the controller is being looked at.