feat(Teleport): disable teleport when controller obscured from view
The ability to disable the teleport if the controller cannot be seen
by the headset has now been added. This means if the controller is
pushed through a wall and the headset cannot have a clear ray cast
to the controller then teleportation is disabled to attempt to prevent
cheating.