feat(Headset): remove need for scripts on headset
  > **Breaking Changes**

Previously, the scripts that required to be on the Headset object would
require placing on the correct object that the headset would finally
become. As SteamVR places the headset on a different game object in
Unity 5.3 to Unity 5.4 and above, it meant that the headset scripts
were cumbersome to use.

This change now just expects any headset related script to be placed
on the `[CameraRig]` prefab as it will look up the headset object
from the DeviceFinder and then apply any logic to the headset
game object rather than needing to directly be on that object.

This change makes it much easier to apply headset scripts as there is
no need to know which object to apply the script to.

The example scenes have been updated to use this new scheme and the
previous `Add_HeadsetCollisionFade` script that was present in some
example scenes has been removed as it no longer requires to do any
clever logic within the editor.