fix(AdaptiveQuality): ensure adaptive quality works on CameraRig
#661 changed the way scripts are added to the `CameraRig` or any of its
children. The problem is that `AdaptiveQuality` uses the `OnPreCull`
method which is only called if the script is added on a game object
that has a `Camera` attached to it. This change fixes the problem by
adding an invocation to the `Camera.onPreCull` delegate and ignoring
all other cameras except the headset camera.