fix(UIPointer): remove UI Pointer artifacts from invalid canvas
If a canvas is set to invalid and `SetWorldCanvas` is called on it
then it will now clean that canvas of any artifacts that is set up
by the UI Pointer when it generates canvases.

This allows a canvas to have an ignore tag set on it at run time or
be taken out of World Space and then cleaned so the UI Pointer now
longer works with it.

Then it can also be enabled again by calling `SetWorldCanvas` again
when the canvas meets the criteria to be a valid canvas.