perf(AdaptiveQuality): change render scale less often
Changing the render scale every two frames is too often because of the
frames the change costs. The changes can lead to frame drops which in
turn signal the system to decrease the render scale even more. This
sometimes results in a "death spiral".

The fix is to decrease the render scale less often and increase it even
less.