feat(Pointer): allow custom transform origin on pointers
The WorldPointer and UIPointer can now both accept a custom transform
to be used as the origin of the pointer (and raycast). This is useful
if the starting position of the pointer and the forward direction
of the pointer is required to be changed. For instance, it's now
possible to rotate the pointers to come out of the controller (or
attached object) at a different angle or from a different position
rather than the centre of the object.