feat(UICanvas): add ui canvas script for defining valid canvas targets
  > **Breaking Changes**

Previously, the UI Pointer was responsible for configuring any
World Space canvases to be compatible with the UI Pointer. However,
this was limiting for a number of reasons:

 * Newly created canvases had to be registered with the UI Pointer
 * Turning off a canvas for interaction was through the UI Pointer
 * Any canvas specific options had to be uniform over all canvases
 * It acted too much like magic having the pointer know how to set
 the canvases up.

The new solution now requires any World Space canvas to have a
UI Canvas script added to it and this will set up the canvas to be
interactable by a UI Pointer. Removing or disabling the UI Canvas
script will reset the canvas back to it's original state and it will
no longer be interactable by the UI Poiinter.

The example scene has been updated to reflect the change and show
how canvases can have independent options now.