fix(Interaction): ensure using state is set correctly after touch/grab
The Using state of an Interactable Object would get out of sync due to
the previous method of stopping an object's use method utilising a
co-routine at the end of the frame.

This was implemented due to StopTouching always being called even when
an object is grabbed. The fix is to remove the Coroutine as it causes
too many issues and to fix the issue at StopTouching by preventing
a stop touch event in the LateUpdate if the object is actually being
grabbed by the touching object, if it's being grabbed then it hasn't
stopped being touched in that instance.

This means that the Using action can get override correctly and no
artifacts are introduced at the end of the frame when stop using was
being called again, creating corruption with the Using State.

The toggle controller visibility has also been fixed by ensuring the
current controller visibility is stored and restored on grab and on
use events.

The `UsingAGrabbedObject` example scene has been updated to provide
more variations of grab and use to aid test interactions.