fix(SDK): prevent crash when requesting non-existant vive render model
The `GetControllerRenderModel` method in the SteamVR SDK can cause a
crash if the scene is not using the default vive controller render
model because the component cannot be retrieved yet the return
statement expects a game object to exist.

The fix is to just check that the found render model exists before
attempting to return the associated game object.