feat(Interaction): add custom secondary grab action to scale objects
A new Custom Secondary Action has been added to the Interactable Object
that takes a new `SecondaryControllerGrabActions` script that defines
the behaviour of what should happen when a second controller grabs
an already grabbed object.

The first Secondary Controller Grab Action to be added is a basic
scale object that will cause the object to scale in the direction the
controller is being moved and it also has parameters on the custom
script to lock certain axis from being scaled.

A new example scene has been added that demonstrates scaling objects.