feat(Interaction): add grab to pointer tip
A new parameter on the World Pointer can now be set that when a
pointer is set to `Interact With Objects = true` rather than on grab
snapping the object to the controller, it now uses the pointer tip
as the attach point.

The length of the pointer is locked at the point of grabbing so it's
not possible to bring the objects closer or move them further away
but a potential feature in the future could be custom public methods
on the pointer scripts that allow the manual entry of a pointer length
so it could be controller via an input of a controller.