fix(SnapDropZone): use highlight container for position
Rather than using the child highlight object, it's better to use the
highlight container for the end transition as this means the actual
highlight area can be different from the destination that the object
will snap into. Whilst, this may not be idea, it's at least better
to offer the option than force it to not be an option.

The position, rotation and size are also forced to the end position
after the co-routine has run to resolve any movement issues during
transition.