fix(Pointer): reset saved beam length if not grabbing
The World Pointer will save the beam length if an object is currently
being grabbed with the pointer and this value was only reset if the
pointer was deactivated, which caused an issue if the pointer was set
to `Always On`. This fix always checks to see if the beam is grabbing
something and if it's not it resets the `savedBeamLength`.