feat(Interaction): remove global touch highlight option
  > ** Breaking Changes **

The Global Touch Highlight Color option on the Interact Touch script
has now been removed as it is somewhat redundant as each Interactable
Object can have a colour set on it individually or multiple objects
can be selected and have the colour set on all of them together.

Having similar options in multiple places causes confusion with what
overrides what.

The Interactable Object also now just has a Touch Highlight Color
option and no longer has a toggle. If there is no highlight colour
required then the colour should be set to `Color.clear` which is
`#00000000` as this will then be ignored when attempting to highlight.