feat(Structure): arrange scripts into sub directory sections
The Scripts directory has been arranged so the scripts that have a
similar responsibility have been grouped together to make it easier
to find relevant scripts and to keep the main scripts directory
from growing larger and more unwieldly.

The Documentation has not been updated in this commit as file changes
would break the simplicity of the git move.