fix(Pointer): ensure controller set up check on base is done in start
The controller set up in the Base Pointer is now done in the Start
method as well as the OnEnable method to ensure it has had enough time
to get the scene set up correctly before looking for the component.

This also fixes an issue when loading between the example scenes.