feat(SDK): split bridge into distinct areas of responsibility
  > ** Breaking Changes **

The SDK Bridge is now split up into four main areas of responsibility:

 * System - core system SDK features.
 * Headset - features for the VR Headset.
 * Controller - features for the VR input device.
 * Boundaries - features for VR play area for room scale.

As these are now split up it would be possible to use different SDKs
to provide support for each of the feature areas making it more
customisable in the future.

For an SDK to be recognised, it now requires a scripting define added
for each SDK area in the Unity `Project Settings -> Player` section.

The following scripting define symbols are available:

 * SteamVR
  * System = `VRTK_SDK_SYSTEM_STEAMVR`
  * Headset = `VRTK_SDK_HEADSET_STEAMVR`
  * Controller = `VRTK_SDK_CONTROLLER_STEAMVR`
  * Boundaries = `VRTK_SDK_BOUNDARIES_STEAMVR`

If these define symbolds are not present then the SDK bridge will
simply fall back to the Base SDK Bridge which provides no functionality
for any of the sections.

A new SDK Manager script has been added that makes it easier to select
a valid SDK and it automatically handles setting up the scripting
symbol defines.

All of the example scenes have been updated to include the SDK Manager
to give better visibility to it's functionality.