feat(PolicyList): add support for layers and rename script
  > ** Breaking Changes **

The TagOrScriptPolicyList has now been renamed simply to Policy List
as it also now allows for checking of object layers and it is better
to simplify the name rather than futher complicate it.

The Policy List also has a custom editor class that allows multiple
selections from the dropdown to choose the most appropraite selection
when checking for inclusion or exclusion.

The relevant scripts that utilise a policy list have also been updated
so their variable names simply refer to a policy list.