feat(PositionRewind): add ability to reset user position on collision
The user's position can now be reset to a last good known position upon
the headset colliding with an object. This can prevent the user from
walking into walls but also not pushing the world around them to
resolve the collisions which can cause sickness.

The Position Rewind can also be given an optional delay to let the user
pull their own head from the collision before the rewind is activated.