feat(SDK): require all sdks to adhere to specific interface
  > ** Breaking Changes **

The SDKs now must adhere to the appropriate interface by implementing
all methods outlined in the relevant interface.

The SteamVR SDKs will now also return back to the Fallback SDK if
the scripting define symbol for SteamVR is not found
(`VRTK_SDK_STEAMVR`).

A spelling error in the SDK_InterfaceController has been amended as
`ControllerElements` was spelt incorrectly, this enum has also been
renamed now to `VRTK_ControllerElements` and has been moved to the
top level of the VRTK namespace.