chore(SDK): remove interface and impose contract with abstract class
  > ** Breaking Changes **

The SDK interfaces have now been removed and the Base classes are now
responsible for enforcing the concrete SDK implemented methods.

All public SDK methods are now abstract in the Base class to ensure
they have to be implemented by any concrete SDK.

The Base classes may also contain protected methods to share common
logic to any concrete implementation but this change now means it
is easier to know which methods need to be implemented when writing
an SDK.

Some of the public enums have been moved into the Base SDK classes
as they make more sense to have them located within the SDK and as
such any references to these enums has been updated to point to the
SDK abstract class.