feat(SDK): add button two on controller
Another button has been added for the controller SDK. The Button Two is
another generic button type that may be used by certain SDKs if the
controller provided has more than one generic button.

The HTC Vive controller only has a Button One (the Application Menu)
but the Oculus Touch has two buttons on each controller (A,B and X,Y)
so this additional button will cater for the Oculus Touch.

Because the HTC Vive doesn't have a Button Two, the SteamVR SDK maps
Button Two to the Application Menu button so if Button Two is set as
an action button, then on the HTC Vive, the Application Menu button
will act as Button Two (as well as Button One).

The Controller Tooltips have also been updated now to include a tooltip
for Button Two and they have also been extended to not show a tooltip
if no text is present for the option (so on the Vive, if no tooltip is
provided for button two, then the tooltip won't be visible).