fix(UICanvas): ensure pre-existing collider/rigidbody is not deleted
The UICanvas script requires a box collider and rigidbody to be created
On Enable to allow for the pointers to correctly interact. When the
UI Canvas is disabled, it cleans up the created collider and rigidbody
by removing them from the GameObject.

However, if the GameObject already has a Box Collider and Rigidbody
then the UI Canvas doesn't automatically create one, it uses the pre
existing ones. But when the script is disabled it deleted the pre
existing components, meaning on re-enable it would just generate the
default components.

This fix ensures that the pre-existing collider and rigidbody are not
deleted when the script is disabled.