feat(layers): expose layers to ignore for all raycasts
Any VRTK script that emits a raycast can now be set to ignore given
layers. Also, the play area object is no longer always forced to the
Ignore Raycast layer if another layer has been selected. If another
layer has been selected then the raycasts in the scripts also need
to be set to ignore that layer as well.