fix(SnapDropZone): ensure object position is right on moving transition
Previously, if the object being dropped into a moving snap drop zone
also had a transition on the drop, then the object would end up not
ending up in the final drop zone place but rather the place the drop
zone was at before the transition started.

This fix ensures that the end position/rotation after the transition
is always matching the current dropzone position.