feat(SDK): add basic support for Oculus Avatar package
The Oculus Avatar package provides access to the Oculus Avatar solution
which contains controller models, hand models and body model.

The VRTK implementation simply checks for the existence of the Oculus
Avatar in the scene and provides custom colliders for the controllers
(only controllers are supported at the moment) and sets up the relevant
transforms for each of the controller elements for use in things like
the controller tooltips.

At the moment, only the Avatar controller models are supported so it's
best off to ensure on the OvrAvatar script that
`Start With Controllers` is ticked and that the `hand_left` and
`hand_right` GameObjects are disabled within the `LocalAvatar`
GameObject (to hide the hand models permanently).

Future updates will look to support the hand models as well.

Some things are not supported, like controller element highlighting
as the Oculus Avatar controller has no child mesh renderers for the
individual button elements and the body element uses a shader with
no colour attribute.