fix(RadialMenu): add option to choose whether to auto generate buttons
Previously, the Radial Menu buttons would always be auto generated in
the Awake method, this may not always be wanted, so now there is an
option of `Generate On Awake` to determine if the buttons should or
should not be generated in the Awake method.

Thanks for @AlbertMN for the original Pull Request with this fix.