fix(SimplePointer): prevent flicker on activate in different position
The simple pointer can flicker when it is activated in a new position
as the position didn't change untit after the renderers were enabled.

This fix disables the renderers on initial beam activation until the
position has been updated and then the renderers are then set to
their correct state.