fix(Interaction): prevent controller becoming visible after use
There was a race condition where if an object's visibility was set
to hide on grab then after it was used it would be automatically
visibile again even if it was still being grabbed.

The code in the Interact Use that was checking to see if the object
should be forced visible again after use should never have been
called as a using object was always required when unuse is executed.

Also, removed a couple of old private methods that no longer get
used in Interact Touch and Interact Use.