fix(Interaction): use local rotation when inverting tracked snap handle
The snap handle rotation is inverted when using a tracked object grab
mechanic. However, it was using the rotation to invert which caused
issues if the object had been rotated in the scene. Rotating the
local rotation of the snap handle is enough to fix the inversion issue
and this does not affect the object if it's world rotation has been
changed.