feat(headset_velocity): add methods to get velocities for headset
The DeviceFinder and SDK classes provide methods to receive the
velocity and angular velocity of controllers. This change adds methods
to all SDKs to retrieve the velocities of the headset. The controller
velocity methods of the VRTK SteamVR SDK have been updated to only
return velocities of controllers, not the headset, because SteamVR uses
indexes for all the tracked objects.