fix(Interaction): parent of the outline highlighter cloned object
Change the parent of the OutlineObjectCopyHighlighter cloned object to
the same parent as the original object.

Change the parent setting to not re-compute world position.

Change the transform assignments from world to local (faster), because
now the parent is the same.

The OutlineObjectCopyHighlighter cloned object needs to have the
parent set to the same parent of the customOutlineModel to exactly
match its shape. Even the PR #828 still causes problems with a hierarchy
of game objects with different scaling (e.g customOutlineModel with its
own scaling as child of the game object with its own scaling, holding
the OutlineObjectCopyHighlighter component).