fix(Interaction): ensure drawer control operates correctly when rotated
  > **Breaking Changes**

The drawer 3D Control now works as expected when it is rotated. It also
correctly snaps shut when it passes a given threshold which has been
changed from a boolean to provide better control over the snap distance
that was previously fixed. The maximum the drawer can be extended is
also now an option on the script.

The Controller Rigidbody Activator has also been updated to allow it
to be disabled on the script without needing to disable the GameObject
as this causes issues with turning off the activator as the Trigger
Exit never gets called.

This now works well with the drawers control as the activator is now
disabled when the drawer is within the snap zone threshold to prevent
unwanted physics glitches.