fix(Interaction): scale of the outline highlighter cloned object
Change the way the transform of the OutlineObjectCopyHighlighter cloned
object is computed.

First set the same parent as the original object, copy the local
transform, then set the parent to `transform`, the owner of the
OutlineObjectCopyHighlighter component, as it was before this commit.

This resolves both PR #828 and PR #850.

The PR #828 still causes problems with a hierarchy of game objects with
different scaling (e.g customOutlineModel with its own scaling as child
of the game object with its own scaling, holding the
OutlineObjectCopyHighlighter component).

The PR #850 broke the Outline Highlighter setting it as the parent,
preventing the outline to follow the object around.