fix(Teleporter): ensure listener coroutine is not run if disabled
The Teleporter script runs a co-routine at the end of the frame to
register any destination markers. However, if the script is disabled
within the first frame then the co-routine is still executed and
registers the listeners even if they're not required.

This fix only registers the listeners if the teleporter script is
enabled and the co-routine is also cancelled when the script is
disabled.