feat(SDK): allow configuration of controller simulator keys
The SDK Simulator controller keys can now be configured via the
simulator camera rig prefab script. It is possible to redefine the
keys used to simulate the controller buttons within the Unity
inspector making it easier to customise.

The access to the raw SDK components in the SDK Bridge have also been
set to public, primarily so this feature can work, but it may also
be required by others if they want direct access to custom functions
on specific SDK components.