fix(Locomotion): enable MoveInPlace engage button change at runtime
The MoveInPlace `Engage Button` could not be changed during runtime
and this fix ensures it can be changed at runtime.

A new method has been added to ControllerEvents to allow easier
registration of button events without needing to do a large case
statement as previously found in the MoveInPlace script.

The getter/setter for the controllers has been removed and replaced
with public variables so they can be toggled at runtime in the
Unity inspector.