fix(Highlighters): provide option to unhighlight on object disable
It is not always desired to unhighlight on disable, in the case of the
snap drop zone highlight objects, they always want to keep their
highlighter active even when the game object is disabled.

The fix is to add a new public parameter that determines whether the
highlighter should disable the highlight effect when the object is
disabled.