feat(Locomotion): add new generic touchpad control script
The new Touchpad Control script is a more generic implementation of the
existing Touchpad Walking script and it allows for custom touchpad
control actions to be written and used in place. This feature makes
the touchpad control aspect more composable and easier to create custom
features.

There is currently only one provided touchpad control action which is
the Slide action and it replicates the same functionality as the
existing Touchpad Walking script.

Because of this duplication in features, the Touchpad Walking script
has now been deprecated and a warning message is shown in the inspector
panel on the script.

The TouchpadWalking example scene has been updated to use the new
Touchpad Control script to provide a demonstration of how to use it.