feat(Locomotion): add axis threshold to snap rotate touchpad control
The Axis Threshold parameter allows the snap rotation to be limited
until the axis reaches the given positive or negative threshold.

This can be utilised to provide a flip rotate if the touchpad is
pulled down (e.g. the axis threshold is set to -0.8) and the rotation
is set to 180 degrees.

This feature also makes the existing Touchpad Movement script
obsolete as all of the functionality can now be replicated by the
Touchpad Control script.