fix(UIPointer): clear existing pointer enter on invalid hover
There was an issue where if a UI button click would disable the canvas
and enable another canvas, then the previous pointer enter selection
would not be cleared and therefore if a button on the new canvas was
selected, it would not have the UI events passed to it as they would
still be tied to the old pointer entry.

This fix ensures the pointer enter event data is cleared up correctly.

Thanks to @nvisionGeo for reporting the issue and posting the fix.