fix(UI): allow non-VR UI usage
This fix makes it possible to use Unity UI in the regular, non-VR way
while making sure the VR UI interactions provided by VRTK still work.

Since this introduces a new class `VRTK_EventSystem` the existing logic
to disable the auto-pausing behavior in Unity 5.5 and above is moved
into the new class. Possible confusion regarding the naming of the
existing `VRTK_EventSystemVRInput` and the new class is resolved by
changing the name of the existing class to `VRTK_VRInputModule`. An
unused constant of `VRTK_UIPointer` is also removed.