fix(Climbing): update example scene to conform with guidelines
The climbing example scene has recently been updated in commit
06eff77f8e236f0440abf8a81cba22e9514499be and the scene causes Unity
to crash as it's become corrupted due to being built in Unity 5.5.

The updated scene has been built in Unity 5.4 to adhere to the support
of the existing and previous release of Unity. The Oculus SDK
gameobjects have also been removed so the scene uses SteamVR by default
as the other example scenes do.

The debug logs have also been removed from the example scripts as they
are not required.