fix(Climbing): ensure play area momentum is kept when ungrabbing
The play area can be in a climbing state but also moving like in the
zip wire example. However, when letting go of a moving climbable
object, the play area would just fall straight down and not keep
any of the momentum from the moving object.

This fix tracks the current velocity of the play area in the Body
Physics script and then applies that momentum as a force to the
play area when the `ApplyBodyVelocity` method is called.

This force is only applied if the optional new `applyMomentum`
parameter is set to true, which it is in the PlayerClimb script.

The magnitude of the existing velocity is taken into consideration
so if an existing velocity is given then the force added is reduced
otherwise the play area has too much force applied.

The `LateUpdate` method in PlayerClimb has been changed back to
an `Update` method as it's not updating the position of the play
area often enough and it causes the play area to just fall with
no momentum being applied even with a force.