fix(Interaction): ensure touchpad axis only reset on touchpad untouch
The touchpad axis was being reset to zero on touchpad untouch but also
if the previous touchpad axis state was the same as the current
touchpad axis state, meaning that every frame the axis would reset to
zero.

This has been fixed by only resetting to zero in the untouch logic.