refactor(ObjectFollow): change attached game object by default
The object follow script is changed so the game object it is attached to
is used for `gameObjectToChange` if none is set. This simplifies usage
of the object follow scripts.

Additionally an issue with the Oculus Avatar SDK is fixed: The OculusVR
boundaries SDK wasn't correctly setting up the transform follow script
to change the game object the OVRAvatar is attached to. This resulted in
the avatar not moving at all.