fix(Internal): remove null coalescing operator usage (??)
It would seem that Unity does some special magic with equality checks
on MonoBehaviour objects meaning that the null coalescing operator
in C# (.Net 2.0) does not work for Unity objects such as GameObjects.

So for example, this won't return what is expectd:

`myvar = myvar ?? othervar;`

should return othervar if myvar is null but it doesn't, instead it
has to be written in long hand like this:

`myvar = myvar != null ? myvar : othervar`

Any reference to `??` has now been removed even if they were
potentially safe just to be sure and to keep things consistent.