feat(Pointer): add method to reset pointer objects at runtime
The pointer cursor and pointer tracer are created in the OnEnable
method which means if a custom tracer/cursor are provided at runtime
then they cannot be switched without disabling and re-enabling the
pointer script.

A new method has been added that can be called which simply destroys
the existing pointer objects and recreates them, so any custom
changes will be present in the update. This also happens without
affecting the existing state of the pointer.