fix(UIPointer): ensure the controller is not null before accessing
The controller parameter could be in a null state yet things would
still attempt to access it causing a critical error.

This fix ensures all calls to the controller are wrapped with a null
check.

The pointerID is now also set in a LateUpdate call every frame rather
than just once in a Coroutine at the start. This means it will update
when the controller ID updates.