fix(Presence): apply new collision tracker to PlayArea for body physics
The Body Physics script was supposed to emit an event on collisions,
however the body physics script is not on the same game object as the
play area collider and rigidbody meaning the script would never emit
the events.

This has been fixed by providing a new Collision Tracker script that
sits on the same game object and simply emits dumb collision events
that the Body Physics script can pick up and forward on.