fix(Interaction): check if touched collider is child of touched object
The check to see if the object is still being touched now uses the
IsChildOf method to check to see if the collider being touched is an
actual relation to the object that has registered being touched.

This is because the game object with the interactable obejct script
on it becomes the Touched Object but this doesn't have to have the
collider on as a child object may contain the collider.

If this is the case then the touch/untouch event is emitted every
frame as the object is failing the `is still being touched` logic.

This fix should resolve that issue.